pyproject.toml
[tool.poetry]
name = "whatsapp-chatbot-smartprints"
version = "0.1.0"
description = ""
authors = ["abdelmalek0 <djamaa.abdelmalek98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
flask = "^3.0.3"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
flask-cors = "^5.0.0"
icecream = "^2.1.3"
langchain-groq = "^0.2.0"
pydub = "^0.25.1"
lxml = "^5.3.0"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-chroma = "^0.1.4"
bs4 = "^0.0.2"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
groq = "^0.11.0"
langchain-core = "^0.3.5"
neuralspace = "^1.2.0"
chromadb = "0.5.0"
langfuse = "^2.51.2"
langdetect = "^1.0.9"
reorder-python-imports = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
